+++ src/data/games/gh5.rs
@@ -121,13 +121,13 @@ pub struct Performance {
     pub score: u64,
 
     /// How many notes were hit successfully.
-    pub notes_hit: u64,
+    pub notes_hit: u32,
 
     /// How many notes were in the chart (TODO: should be const across different scores of the chart)
-    pub notes_total: u64,
+    pub notes_total: u32,
 
     /// The maximum streak achieved during the performance.
-    pub max_streak: u64,
+    pub max_streak: u32,
 }
 
 #[typetag::serde(name = "gh5")]
